fix: BTS-226 / 스터디룸 페이지 접근 에러 해결
